+++ components/faphub/catalogtab/api/src/main/java/com/flipperdevices/faphub/catalogtab/api/CatalogTabApi.kt
@@ -2,12 +2,12 @@ package com.flipperdevices.faphub.catalogtab.api
 
 import androidx.compose.runtime.Composable
 import com.flipperdevices.faphub.dao.api.model.FapCategory
-import com.flipperdevices.faphub.dao.api.model.FapItem
+import com.flipperdevices.faphub.dao.api.model.FapItemShort
 
 interface CatalogTabApi {
     @Composable
     fun ComposableCatalogTab(
-        onOpenFapItem: (FapItem) -> Unit,
+        onOpenFapItem: (FapItemShort) -> Unit,
         onCategoryClick: (FapCategory) -> Unit
     )
 }