+++ components/faphub/catalogtab/impl/src/main/java/com/flipperdevices/faphub/catalogtab/impl/api/CatalogTabApiImpl.kt
@@ -5,7 +5,7 @@ import com.flipperdevices.core.di.AppGraph
 import com.flipperdevices.faphub.catalogtab.api.CatalogTabApi
 import com.flipperdevices.faphub.catalogtab.impl.composable.ComposableCatalogTabScreen
 import com.flipperdevices.faphub.dao.api.model.FapCategory
-import com.flipperdevices.faphub.dao.api.model.FapItem
+import com.flipperdevices.faphub.dao.api.model.FapItemShort
 import com.flipperdevices.faphub.installation.api.FapInstallationUIApi
 import com.squareup.anvil.annotations.ContributesBinding
 import javax.inject.Inject
@@ -16,7 +16,7 @@ class CatalogTabApiImpl @Inject constructor(
 ) : CatalogTabApi {
     @Composable
     override fun ComposableCatalogTab(
-        onOpenFapItem: (FapItem) -> Unit,
+        onOpenFapItem: (FapItemShort) -> Unit,
         onCategoryClick: (FapCategory) -> Unit
     ) {
         ComposableCatalogTabScreen(
@@ -24,7 +24,7 @@ class CatalogTabApiImpl @Inject constructor(
             onCategoryClick = onCategoryClick,
             installationButton = { fapItem, modifier, fontSize ->
                 fapInstallationUIApi.ComposableButton(
-                    fapItem = fapItem,
+                    fapItemId = fapItem?.id,
                     modifier = modifier,
                     textSize = fontSize
                 )