+++ components/faphub/catalogtab/impl/src/main/java/com/flipperdevices/faphub/catalogtab/impl/composable/ComposableCatalogTab.kt
@@ -15,15 +15,15 @@ import com.flipperdevices.faphub.catalogtab.impl.composable.categories.Composabl
 import com.flipperdevices.faphub.catalogtab.impl.viewmodel.CategoriesViewModel
 import com.flipperdevices.faphub.catalogtab.impl.viewmodel.FapsListViewModel
 import com.flipperdevices.faphub.dao.api.model.FapCategory
-import com.flipperdevices.faphub.dao.api.model.FapItem
+import com.flipperdevices.faphub.dao.api.model.FapItemShort
 import tangle.viewmodel.compose.tangleViewModel
 
 @Composable
 fun ComposableCatalogTabScreen(
-    onOpenFapItem: (FapItem) -> Unit,
+    onOpenFapItem: (FapItemShort) -> Unit,
     onCategoryClick: (FapCategory) -> Unit,
     modifier: Modifier = Modifier,
-    installationButton: @Composable (FapItem?, Modifier, TextUnit) -> Unit
+    installationButton: @Composable (FapItemShort?, Modifier, TextUnit) -> Unit
 ) {
     val fapsListViewModel = tangleViewModel<FapsListViewModel>()
     val fapsList = fapsListViewModel.faps.collectAsLazyPagingItems()