+++ components/faphub/category/impl/src/main/java/com/flipperdevices/faphub/category/impl/api/FapHubCategoryApiImpl.kt
@@ -52,7 +52,7 @@ class FapHubCategoryApiImpl @Inject constructor(
                 },
                 installationButton = { fapItem, modifier, fontSize ->
                     fapInstallationUIApi.ComposableButton(
-                        fapItem = fapItem,
+                        fapItemId = fapItem?.id,
                         modifier = modifier,
                         textSize = fontSize
                     )