+++ components/faphub/category/impl/src/main/java/com/flipperdevices/faphub/category/impl/composable/ComposableFapHubCategoryScreen.kt
@@ -12,7 +12,7 @@ import com.flipperdevices.core.ui.ktx.OrangeAppBarWithIcon
 import com.flipperdevices.faphub.appcard.composable.paging.ComposableFapsList
 import com.flipperdevices.faphub.appcard.composable.paging.ComposableSortChoice
 import com.flipperdevices.faphub.category.impl.viewmodel.FapHubCategoryViewModel
-import com.flipperdevices.faphub.dao.api.model.FapItem
+import com.flipperdevices.faphub.dao.api.model.FapItemShort
 import tangle.viewmodel.compose.tangleViewModel
 import com.flipperdevices.core.ui.res.R as DesignSystem
 
@@ -20,9 +20,9 @@ import com.flipperdevices.core.ui.res.R as DesignSystem
 fun ComposableFapHubCategory(
     onBack: () -> Unit,
     onOpenSearch: () -> Unit,
-    onOpenFapItem: (FapItem) -> Unit,
+    onOpenFapItem: (FapItemShort) -> Unit,
     modifier: Modifier = Modifier,
-    installationButton: @Composable (FapItem?, Modifier, TextUnit) -> Unit
+    installationButton: @Composable (FapItemShort?, Modifier, TextUnit) -> Unit
 ) {
     val categoryViewModel = tangleViewModel<FapHubCategoryViewModel>()
     val fapsList = categoryViewModel.faps.collectAsLazyPagingItems()