+++ components/faphub/dao/api/src/main/java/com/flipperdevices/faphub/dao/api/FapNetworkApi.kt
@@ -2,16 +2,24 @@ package com.flipperdevices.faphub.dao.api
 
 import com.flipperdevices.faphub.dao.api.model.FapCategory
 import com.flipperdevices.faphub.dao.api.model.FapItem
+import com.flipperdevices.faphub.dao.api.model.FapItemShort
 import com.flipperdevices.faphub.dao.api.model.SortType
 
 interface FapNetworkApi {
-    suspend fun getFeaturedItem(): FapItem
+    suspend fun getFeaturedItem(): FapItemShort
     suspend fun getAllItem(
         category: FapCategory? = null,
-        sortType: SortType
-    ): List<FapItem>
+        sortType: SortType,
+        offset: Int,
+        limit: Int
+    ): List<FapItemShort>
 
-    suspend fun search(query: String): List<FapItem>
+    suspend fun search(
+        query: String,
+        offset: Int,
+        limit: Int
+    ): List<FapItemShort>
+
     suspend fun getCategories(): List<FapCategory>
     suspend fun getFapItemById(id: String): FapItem
 }