+++ components/faphub/dao/network/src/main/java/com/flipperdevices/faphub/dao/network/model/ApplicationVersionShort.kt
@@ -1,0 +1,10 @@
+package com.flipperdevices.faphub.dao.network.model
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ApplicationVersionShort(
+    @SerialName("_id") val id: String,
+    @SerialName("version") val version: String,
+)