+++ components/faphub/dao/network/src/main/java/com/flipperdevices/faphub/dao/network/retrofit/utils/DateSerializer.kt
@@ -1,0 +1,31 @@
+package com.flipperdevices.faphub.dao.network.retrofit.utils
+
+import android.annotation.SuppressLint
+import kotlinx.serialization.KSerializer
+import kotlinx.serialization.Serializer
+import kotlinx.serialization.descriptors.PrimitiveKind
+import kotlinx.serialization.descriptors.PrimitiveSerialDescriptor
+import kotlinx.serialization.descriptors.SerialDescriptor
+import kotlinx.serialization.encoding.Decoder
+import kotlinx.serialization.encoding.Encoder
+import java.text.DateFormat
+import java.text.SimpleDateFormat
+import java.util.Date
+
+@Serializer(forClass = Date::class)
+object DateSerializer : KSerializer<Date> {
+    override val descriptor: SerialDescriptor =
+        PrimitiveSerialDescriptor("Date", PrimitiveKind.STRING)
+
+    @SuppressLint("SimpleDateFormat")
+    private val df: DateFormat = SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSSZ")
+
+    override fun deserialize(decoder: Decoder): Date {
+        val rawDate = decoder.decodeString()
+        return df.parse(rawDate) ?: error("Failed parse $rawDate")
+    }
+
+    override fun serialize(encoder: Encoder, value: Date) {
+        encoder.encodeString(df.format(value))
+    }
+}