+++ components/faphub/installation/api/src/main/java/com/flipperdevices/faphub/installation/api/FapInstallationUIApi.kt
@@ -3,12 +3,11 @@ package com.flipperdevices.faphub.installation.api
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.TextUnit
-import com.flipperdevices.faphub.dao.api.model.FapItem
 
 interface FapInstallationUIApi {
     @Composable
     fun ComposableButton(
-        fapItem: FapItem?,
+        fapItemId: String?,
         modifier: Modifier,
         textSize: TextUnit
     )