+++ components/faphub/search/impl/src/main/java/com/flipperdevices/faphub/search/impl/viewmodel/FapsSearchPagingSource.kt
@@ -1,29 +1,16 @@
 package com.flipperdevices.faphub.search.impl.viewmodel
 
-import androidx.paging.PagingSource
-import androidx.paging.PagingState
+import com.flipperdevices.core.pager.OffsetAndLimitPagingSource
 import com.flipperdevices.faphub.dao.api.FapNetworkApi
-import com.flipperdevices.faphub.dao.api.model.FapItem
+import com.flipperdevices.faphub.dao.api.model.FapItemShort
+
+internal const val FAPS_PAGE_SIZE = 100
 
 class FapsSearchPagingSource(
     private val fapNetworkApi: FapNetworkApi,
     private val searchRequest: String
-) : PagingSource<Int, FapItem>() {
-    override fun getRefreshKey(state: PagingState<Int, FapItem>): Int? {
-        return state.anchorPosition
-    }
-
-    override suspend fun load(params: LoadParams<Int>): LoadResult<Int, FapItem> {
-        return try {
-            val nextPage = params.key ?: 1
-            val faps = fapNetworkApi.search(searchRequest)
-            LoadResult.Page(
-                data = faps,
-                prevKey = if (nextPage == 1) null else nextPage - 1,
-                nextKey = nextPage + 1
-            )
-        } catch (exception: Throwable) {
-            return LoadResult.Error(exception)
-        }
+) : OffsetAndLimitPagingSource<FapItemShort>(FAPS_PAGE_SIZE) {
+    override suspend fun load(offset: Int, limit: Int): List<FapItemShort> {
+        return fapNetworkApi.search(searchRequest, offset, limit)
     }
 }